+++ scilab/modules/hdf5/sci_gateway/cpp/sci_hdf5_listvar_v3.cpp
@@ -738,7 +738,11 @@ static bool read_struct(int dataset, VarInfo6& info)
             //import field
             for (int j = 0; j < refcount; ++j)
             {
-                int data = H5Rdereference(refs, H5R_OBJECT, &vrefs[j]);
+                int data = H5Rdereference(refs,
+#if H5_VERSION_GE(1,10,0)
+                                          H5P_DATASET_ACCESS_DEFAULT,
+#endif
+                                          H5R_OBJECT, &vrefs[j]);
                 if (data < 0)
                 {
                     return false;